# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gestión de Garantías",
    "version": "1.0.0",
    "description": "Documentación de la API para la gestión de garantías."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor local alternativo"
    },
    {
      "url": "https://tu-dominio-vercel.vercel.app",
      "description": "Servidor de producción"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Iniciar sesión",
        "description": "Autenticar a un usuario y devolver un token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin123"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas."
          }
        }
      }
    },
    "/api/warranties": {
      "get": {
        "summary": "Obtener garantías",
        "description": "Devuelve una lista de garantías registradas.",
        "responses": {
          "200": {
            "description": "Lista de garantías obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "name": { "type": "string" },
                      "status": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}